nonebot_tools/nonebot-plugin-ncm/requirements.txt
aiofiles
nonebot2>=2.0.0b1
httpx
tinydb
pydantic
pyncm==1.6.4.0